+++ src/main/java/com/haubigdata24_12/agriculturalmachinery/controller/payment/method/AlipayPayment.java
@@ -4,8 +4,8 @@ package com.haubigdata24_12.agriculturalmachinery.controller.payment.method;
 import com.haubigdata24_12.agriculturalmachinery.model.PaymentInfo;
 
 public class AlipayPayment extends PaymentInfo {
-    private int accountNo;
-    public AlipayPayment(double paymentAmount, int accountNo) {
+    private long accountNo;
+    public AlipayPayment(double paymentAmount, long accountNo) {
         super(paymentAmount);
 
         this.accountNo = accountNo;
