+++ src/main/java/com/haubigdata24_12/agriculturalmachinery/controller/payment/method/BankCardPayment.java
@@ -3,8 +3,8 @@ package com.haubigdata24_12.agriculturalmachinery.controller.payment.method;
 import com.haubigdata24_12.agriculturalmachinery.model.PaymentInfo;
 
 public class BankCardPayment extends PaymentInfo {
-    private int cardNo;
-    public BankCardPayment(double paymentAmount, int cardNo) {
+    private long cardNo;
+    public BankCardPayment(double paymentAmount, long cardNo) {
         super(paymentAmount);
 
         this.cardNo = cardNo;
