+++ src/main/java/com/haubigdata24_12/agriculturalmachinery/controller/payment/method/WechatPayment.java
@@ -3,8 +3,8 @@ package com.haubigdata24_12.agriculturalmachinery.controller.payment.method;
 import com.haubigdata24_12.agriculturalmachinery.model.PaymentInfo;
 
 public class WechatPayment extends PaymentInfo {
-    private int OpenId;
-    public WechatPayment(double paymentAmount ,int OpenId) {
+    private long OpenId;
+    public WechatPayment(double paymentAmount ,long OpenId) {
         super(paymentAmount);
 
         this.OpenId=OpenId;
